// NextSpring/frontend/src/lib/backend/apiV1/schema.d.ts
export interface components {
    schemas: {
// ...
        PostWithContentDto: {
            /** Format: int64 */
            id: number;
            /** Format: date-time */
            createDate: string;
            /** Format: date-time */
            modifyDate: string;
            /** Format: int64 */
            authorId: number;
            authorName: string;
            title: string;
            published: boolean;
            listed: boolean;
            content: string;
            actorCanModify?: boolean;
            actorCanDelete?: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}